+++ CCU_validation_scoring/score_submission.py
@@ -100,7 +100,7 @@ def score_nd_submission_dir_cli(args):
 		merge_ref_time_gap = None
 
 	ensure_output_dir(args.output_dir)
-	ref = preprocess_reference_dir(ref_dir = args.reference_dir, scoring_index = scoring_index, task = "norms", text_gap = merge_ref_text_gap, time_gap = merge_ref_time_gap, merge_label = args.merge_ref_label, dump_inputs=args.dump_inputs, output_dir=args.output_dir)
+	ref = preprocess_reference_dir(ref_dir = args.reference_dir, scoring_index = scoring_index, task = "norms", text_gap = merge_ref_text_gap, time_gap = merge_ref_time_gap, merge_label = args.merge_ref_label, dump_inputs=args.dump_inputs, output_dir=args.output_dir, fix_ref_status_conflict_label=args.fix_ref_status_conflict)
 
 	if args.norm_list_file:
 		ref = process_subset_norm_emotion(args.norm_list_file, ref)