+++ scripts/generate_perfect_submission.py
@@ -47,7 +47,7 @@ def generate_submission_dir(output_dir, task):
 
 	return output_submission
 
-def generate_perfect_submission(task, reference_dir, scoring_index_file, output_dir, text_gap, time_gap, merge_label):
+def generate_perfect_submission(task, submission_format, reference_dir, scoring_index_file, output_dir, text_gap, time_gap, merge_label, fix_ref_status_conflict = None):
 
 	try:
 		scoring_index = pd.read_csv(scoring_index_file, usecols = ['file_id'], sep = "\t")
@@ -55,8 +55,9 @@ def generate_perfect_submission(task, reference_dir, scoring_index_file, output_
 		logger.error('ERROR:GENERATING:{} is not a valid scoring index file'.format(scoring_index_file))
 		exit(1)
 
-	ref = preprocess_reference_dir(reference_dir, scoring_index, task, text_gap, time_gap, merge_label)
-	ref.loc[ref["status"] == "adhere,violate", "status"] = "adhere"
+	ref = preprocess_reference_dir(reference_dir, scoring_index, task, text_gap, time_gap, merge_label, fix_ref_status_conflict_label=fix_ref_status_conflict)
+	if (task == 'norms'):
+		ref.loc[ref["status"] == "adhere,violate", "status"] = "adhere"
 
 	output_submission = generate_submission_dir(output_dir, task)
 
@@ -71,7 +72,15 @@ def generate_perfect_submission(task, reference_dir, scoring_index_file, output_
 		
 		sub_ref = ref.loc[ref["file_id"] == i]
 		sub_ref_ann = sub_ref.loc[sub_ref["Class"] != "noann"]
-		filter_sub_ref_ann = sub_ref_ann[["file_id","Class","start","end","status"]]
+		if (task == 'norms'):
+			if (submission_format == "open"):
+				segment_df = pd.read_csv(os.path.join(reference_dir,"docs","segments.tab"), sep = "\t")
+				sub_ref_ann = sub_ref_ann.merge(segment_df, on=["file_id","start","end"])
+				filter_sub_ref_ann = sub_ref_ann[["file_id","segment_id","Class","status"]]
+			else:
+				filter_sub_ref_ann = sub_ref_ann[["file_id","Class","start","end","status"]]
+		else:
+			filter_sub_ref_ann = sub_ref_ann[["file_id","Class","start","end"]]
 		rename_filter_sub_ref_ann = filter_sub_ref_ann.rename(columns = {"Class": task_column})
 		rename_filter_sub_ref_ann["llr"] = 0.5  
 		index_df = generate_submission_file(i, rename_filter_sub_ref_ann, output_submission, index_df, "True")
@@ -80,10 +89,12 @@ def generate_perfect_submission(task, reference_dir, scoring_index_file, output_
 def main():
 	parser = argparse.ArgumentParser(description='Generate a random norm/emotion submission')
 	parser.add_argument('-t', '--task', choices=['norms', 'emotions'], required=True, help = 'norms, emotions')
+	parser.add_argument('-sf','--submission-format', type=str, default="regular", choices=['regular','open'], help='choose submission format, regular is for CCU evaluation while open is for OpenCCU')
 	parser.add_argument('-ref','--reference-dir', type=str, required=True, help='Reference directory')
 	parser.add_argument("-xR", "--merge_ref_text_gap", type=str, required=False, help="merge reference text gap character")
 	parser.add_argument("-aR", "--merge_ref_time_gap", type=str, required=False, help="merge reference time gap second")
 	parser.add_argument("-vR", "--merge_ref_label", type=str, choices=['class', 'class-status'], required=False, help="choose class or class-status to define how to handle the adhere/violate labels for the reference norm instances merging. class is to use the class label only (ignoring status) to merge and class-status is to use the class and status label to merge")
+	parser.add_argument("-f", "--fix_ref_status_conflict", action='store_true', help="set reference annotation to noann when there are the same annotations but different status")
 	parser.add_argument('-i','--scoring-index-file', type=str, required=True, help='Use to filter file from scoring (REF)')
 	parser.add_argument("-o", "--output_dir", type=str, required=True, help="Output directory")
 
@@ -99,7 +110,7 @@ def main():
 	else:
 		merge_ref_time_gap = None
 
-	generate_perfect_submission(args.task, args.reference_dir, args.scoring_index_file, args.output_dir, merge_ref_text_gap, merge_ref_time_gap, args.merge_ref_label)
+	generate_perfect_submission(args.task, args.submission_format, args.reference_dir, args.scoring_index_file, args.output_dir, merge_ref_text_gap, merge_ref_time_gap, args.merge_ref_label, args.fix_ref_status_conflict)
 
 if __name__ == '__main__':
 	main()